creacion de routa para recuperar postbyId

+++ models/post.model.js
@@ -10,7 +10,7 @@ const create = ({ date, idTypePost, text, usuario_id }) => {
     return executeQuery('INSERT INTO post (date, idTypePost, text, usuario_id) VALUES (?, ?, ?, ?)', [date, idTypePost, text,usuario_id]);
 }
 const getById = (postId) => {
-    return executeQuery('SELECT * FROM post WHERE usuario_id = ?', [postId]);
+    return executeQuery('SELECT * FROM post WHERE id = ?', [postId]);
 }
 
 const deleteById = (postId) => {
@@ -18,12 +18,17 @@ const deleteById = (postId) => {
 }
 
 
+const getByUserId = (userId) => {
+    return executeQuery('SELECT * FROM post WHERE usuario_id = ?', [userId]);
+}
 
 
 
 
 
-module.exports = { create, getAll,getById,deleteById };
+
+
+module.exports = { create, getAll,getById,deleteById, getByUserId};
